truck-map-dashboard/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
numpy==1.24.3
Pillow==10.0.1
eventlet==0.33.3
python-socketio==5.9.0
paho-mqtt==1.6.1